requirements.txt
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
datasets==2.14.0
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.1
Pillow==10.0.1
matplotlib==3.8.1
seaborn==0.13.0
tqdm==4.66.1
accelerate==0.24.1
bitsandbytes==0.41.1
peft==0.7.0
clip-by-openai==0.0.0.dev0
open-clip-torch==2.20.0
scipy==1.11.3
opencv-python==4.8.1.78
Pillow==10.0.1
streamlit==1.28.1
plotly==5.18.0
wandb==0.16.0
python-dotenv==1.0.0
